app/backend/requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
django-cors-headers
python-dotenv
celery
django-celery-results
